Re-define Class Hierarchy
- Moves ModInfo base class of namespace
- Methods moved to own class that inherits from Updater
